requirement.txt
Package                 Version
----------------------- ------------------
absl-py                 2.1.0
addict                  2.4.0
albucore                0.0.17
albumentations          1.3.0
annotated-types         0.7.0
asttokens               2.4.1
augly                   1.0.0
Brotli                  1.0.9
certifi                 2024.8.30
charset-normalizer      3.3.2
classy-vision           0.7.0
colorama                0.4.6
comm                    0.2.2
contourpy               1.3.0
cycler                  0.12.1
debugpy                 1.6.7
decorator               5.1.1
eval_type_backport      0.2.0
exceptiongroup          1.2.2
executing               2.1.0
faiss-cpu               1.9.0
filelock                3.13.1
fonttools               4.54.1
fsspec                  2024.9.0
fvcore                  0.1.5.post20221221
grpcio                  1.62.2
huggingface-hub         0.25.1
idna                    3.7
imageio                 2.35.1
importlib_metadata      8.5.0
iopath                  0.1.10
ipykernel               6.29.5
ipython                 8.27.0
jedi                    0.19.1
Jinja2                  3.1.4
joblib                  1.4.2
jupyter_client          8.6.3
jupyter_core            5.7.2
kiwisolver              1.4.7
lazy_loader             0.4
llvmlite                0.43.0
lmdb                    1.5.1
loguru                  0.7.2
Markdown                3.4.1
markdown-it-py          3.0.0
MarkupSafe              2.1.3
matplotlib              3.9.2
matplotlib-inline       0.1.7
mdurl                   0.1.2
mkl_fft                 1.3.10
mkl_random              1.2.7
mkl-service             2.4.0
mmcv                    2.2.0
mmcv-full               1.7.2
mmengine                0.10.5
mpmath                  1.3.0
nest_asyncio            1.6.0
networkx                3.2.1
numba                   0.60.0
numpy                   1.26.4
opencv-python           4.10.0
opencv-python-headless  4.10.0
packaging               24.1
pandas                  2.2.3
parso                   0.8.4
pickleshare             0.7.5
pillow                  10.4.0
pip                     24.2
platformdirs            4.3.6
portalocker             2.10.1
prompt_toolkit          3.0.47
protobuf                4.25.3
psutil                  5.9.0
pure_eval               0.2.3
pydantic                2.9.2
pydantic_core           2.23.4
Pygments                2.18.0
pyparsing               3.1.4
PySnooper               1.2.1
PySocks                 1.7.1
python-dateutil         2.9.0
python-magic-bin        0.4.14
pytz                    2024.2
pywin32                 305.1
PyYAML                  6.0.1
pyzmq                   25.1.2
qudida                  0.0.4
regex                   2024.9.11
requests                2.32.3
rich                    13.8.1
safetensors             0.4.5
scikit-image            0.24.0
scikit-learn            1.5.1
scipy                   1.13.1
setuptools              75.1.0
six                     1.16.0
stack-data              0.6.2
sympy                   1.13.2
tabulate                0.9.0
tensorboard             2.17.1
tensorboard_data_server 0.7.0
termcolor               2.4.0
threadpoolctl           3.5.0
tifffile                2024.9.20
timm                    0.6.12
tokenizers              0.20.0
tomli                   2.0.1
torch                   2.4.1+cu118
torchaudio              2.4.1
TorchSnooper            0.8
torchvision             0.19.1
tornado                 6.4.1
tqdm                    4.66.5
traitlets               5.14.3
transformers            4.45.2
tslearn                 0.6.3
typing_extensions       4.11.0
tzdata                  2024.2
urllib3                 2.2.2
wcwidth                 0.2.13
Werkzeug                3.0.3
wheel                   0.44.0
win-inet-pton           1.1.0
win32-setctime          1.1.0
yacs                    0.1.8
yapf                    0.40.2
zipp                    3.20.2